iox-#1391 Move 'error_reporting' to new location
