+++ .clang-tidy-diff-scans.txt
@@ -18,8 +18,6 @@
 
 ./iceoryx_hoofs/test/moduletests/test_posix*
 
-./iceoryx_hoofs/include/iceoryx_hoofs/error_reporting/**/*
-./iceoryx_hoofs/source/error_reporting/**/*
 ./iceoryx_hoofs/testing/include/iceoryx_hoofs/testing/error_reporting/*
 ./iceoryx_hoofs/testing/error_reporting/*
