+++ iceoryx_hoofs/reporting/source/console_logger.cpp
