+++ iceoryx_hoofs/reporting/source/logger.cpp
