+++ iceoryx_hoofs/test/moduletests/error_reporting/test_auxiliary_types.cpp
@@ -17,7 +17,7 @@
 #include "test.hpp"
 #include <gtest/gtest.h>
 
-#include "iceoryx_hoofs/error_reporting/types.hpp"
+#include "iox/error_reporting/types.hpp"
 
 namespace
 {
