+++ iceoryx_hoofs/vocabulary/include/iox/string.hpp
@@ -17,7 +17,7 @@
 #ifndef IOX_HOOFS_VOCABULARY_STRING_HPP
 #define IOX_HOOFS_VOCABULARY_STRING_HPP
 
-#include "iceoryx_hoofs/error_reporting/error_reporting_macros.hpp"
+#include "iox/assertions.hpp"
 #include "iox/buffer_info.hpp"
 #include "iox/detail/string_internal.hpp"
 #include "iox/detail/string_type_traits.hpp"
